+++ Example/Sources/Data Generation/SampleData.swift
@@ -46,7 +46,7 @@ final internal class SampleData {
     
     let messageImages: [UIImage] = [#imageLiteral(resourceName: "img1"), #imageLiteral(resourceName: "img2")]
 
-    let messageTypes = ["Text", "Text", "Text", "AttributedText", "Location", "Photo", "Emoji", "Video", "URL", "Phone", "Custom"]
+    let messageTypes = ["Text", "Text", "Text", "AttributedText", "Location", "Photo", "Emoji", "Video", "URL", "Phone", "Mention", "Hashtag", "Custom"]
 
     let emojis = [
         "👍",
@@ -150,6 +150,10 @@ final internal class SampleData {
             return MockMessage(text: "https://github.com/MessageKit", sender: sender, messageId: uniqueID, date: date)
         case "Phone":
             return MockMessage(text: "123-456-7890", sender: sender, messageId: uniqueID, date: date)
+        case "Mention":
+            return MockMessage(text: "@messagekit", sender: sender, messageId: uniqueID, date: date)
+        case "Hashtag":
+            return MockMessage(text: "#messagekit", sender: sender, messageId: uniqueID, date: date)
         case "Custom":
             return MockMessage(custom: "Someone left the conversation", sender: system, messageId: uniqueID, date: date)
         default: